move github and ml to their own files
