+++ detection_rules/__init__.py
@@ -12,9 +12,11 @@ from . import (  # noqa: E402
     devtools,
     docs,
     eswrap,
+    ghwrap,
     kbwrap,
     main,
     mappings,
+    ml,
     misc,
     rule_formatter,
     rule_loader,
@@ -26,10 +28,12 @@ __all__ = (
     'devtools',
     'docs',
     'eswrap',
+    'ghwrap',
     'kbwrap',
     'mappings',
     "main",
     'misc',
+    'ml',
     'rule_formatter',
     'rule_loader',
     'schemas',
